requirements.txt
Flask==2.1.2
requests==2.28.1
python-telegram-bot==13.12
